[refactor](player) Extract player-related logic to `PlayerService` for better modularity and maintainability
