+++ composeApp/src/commonMain/kotlin/world/hachimi/app/ui/search/SearchScreen.kt
@@ -82,7 +82,7 @@ fun SearchScreen(
             key = { _, item -> item.displayId }
         ) { index, item ->
             SearchSongItem(item, Modifier.widthIn(min = 300.dp).padding(vertical = 12.dp), {
-                global.insertToQueue(item.displayId, true, false)
+                global.player.insertToQueue(item.displayId, true, false)
             })
         }
     }
